Thanks to all that made CM7 for the NT possible
Special thanks go to Bauwks for his 2nd boot without which no custom rom wodul be possible.
ANotehr big thank you to Fattire for his work with Nook Color which is where we started to build it for the NT and for all eth help he selflessly gave whenever we asked for it.

Also thanks to anybody who tetsed or helped out in anyway at all, you all made this release :)

Further thanks can be foudn in our thread on XDA in rom thread under Nook Tablet development.